playground/online-learning-game/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.70.0
google-cloud-storage>=2.18.0
google-cloud-firestore>=2.19.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
